refactor: Move ActivitiesApiService to remote package
