+++ core/data/src/main/java/com/projectlab/core/data/repository/ActivitiesRepositoryImpl.kt
@@ -5,7 +5,7 @@ import com.projectlab.core.data.mapper.toDomain
 import com.projectlab.core.domain.repository.ActivitiesRepository
 import javax.inject.Inject
 import com.projectlab.core.domain.model.Activity
-import com.projectlab.core.data.mapper.toDomainList
+import com.projectlab.core.data.remote.ActivitiesApiService
 import com.projectlab.core.domain.repository.TokenProvider
 import com.projectlab.core.domain.util.DataError
 import com.projectlab.core.domain.util.Result
@@ -20,15 +20,15 @@ class ActivitiesRepositoryImpl @Inject constructor(
         lon: Double
     ): Result<List<Activity>, DataError.Network> {
         val token = tokenProvider.getAccessToken()
-        Log.d("ActivitiesRepository", "Obtained token: $token")
         if (token.isEmpty()) {
             throw Exception("Access token is missing")
         }
         return try {
             val response = apiService.getActivitiesByLocation(latitude = lat, longitude = lon)
+            response.data
             Result.Success(response.data.map { it.toDomain() })
         } catch (e: Exception) {
-            Log.e("ActivitiesRepository", "Error fetching activities", e)
+            e.printStackTrace()
             Result.Error(DataError.Network.UNKNOWN)
         }
     }
